requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<6.0
